backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
python-multipart==0.0.20
rapidfuzz>=3.9.0
openai>=1.30.0
anthropic>=0.40.0
python-dotenv>=1.0.0
